pyproject.toml
[project]
name = "arena"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # MCP server + client
    "mcp>=1.0",
    # Frameworks
    "anthropic>=0.40",
    "strands-agents>=0.1",
    "strands-agents-mcp>=0.1",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-mcp-adapters>=0.1",
    "langgraph>=0.2",
    "langchain-mcp-adapters>=0.1",
    "crewai>=0.80",
    "google-adk>=0.3",
    # Metrics
    "radon>=6.0",
    # Utilities
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
